Convert is_not_terrain_object function to use generics
- Remove unnecessary type casting

+++ src/game/constants.rs
@@ -38,7 +38,7 @@ pub const AMMO_POSITIONS: [[f32; 2]; 4] = [[-640.0, -560.0], [-700.0, 400.0], [7
 pub const HOUSE_POSITIONS: [[f32; 2]; 2] = [[-36.0, 644.0], [506.0, 230.0]];
 pub const TREE_POSITIONS: [[f32; 2]; 5] = [[-506.0, -230.0], [368.0, -368.0], [-690.0, -506.0], [-874.0, -92.0], [-690.0, 138.0]];
 
-pub const TERRAIN_OBJECTS: [[usize; 2]; 34] = [ [ 48, 50 ],
+pub const TERRAIN_OBJECTS: [[i32; 2]; 34] = [ [ 48, 50 ],
   [ 48, 51 ],
   [ 48, 52 ],
   [ 49, 50 ],
